requirements.txt
Flask
flask-cors
PyYAML
Flask-RESTful